% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numbered_labels.R
\name{alpha_to_int}
\alias{alpha_to_int}
\title{Convert an alphabetical character object from A to Z into a corresponding
integer}
\usage{
alpha_to_int(x, dict = LETTERS, n = 1, quiet = TRUE, call = parent.frame())
}
\arguments{
\item{x}{Character vector of length n strings to compare to dict. Typically,
letters from "A" to "Z". Case sensitive.}

\item{dict}{Character vector to match to x. Default: LETTERS.}

\item{n}{Maximum character length for non-NA objects permitted. Set to NULL
or >1 if dict includes objects with more than one character.}

\item{quiet}{If \code{TRUE}, suppress warnings for introduction of NA values
through coercion.}

\item{call}{Default: \code{\link[=parent.frame]{parent.frame()}}. Passed to input checking functions
to improve error messages.}
}
\value{
A length 1 integer between 1 and 26.
}
\description{
Integers and NA values are passed as is. Double or characters with no
corresponding Roman numeral are converting to NA values.
}
