% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cardinal_bearing.R
\name{as_cardinal_bearing}
\alias{as_cardinal_bearing}
\title{Convert a numeric bearing value to the closest cardinal bearing}
\usage{
as_cardinal_bearing(x, winds = 8, cols = c("bearing", "cardinal_bearing"))
}
\arguments{
\item{x}{A numeric vector with degrees or a data.frame with column name
matching the first name in cols.}

\item{winds}{Number of winds to use for results (4, 8, or 16).}

\item{cols}{A length 2 character vector where the first value is a column
name containing bearing values and the second is the name of the new column
added to the data.frame. Required if x is a data.frame.}
}
\value{
A named numeric vector with cardinal bearings (and wind names) or a
data.frame with an added column containing the cardinal bearings.
}
\description{
Convert a numeric bearing value to the closest cardinal bearing
}
