% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cardinal_degrees.R
\name{as_cardinal_degrees}
\alias{as_cardinal_degrees}
\title{What are the closest cardinal degrees to a numeric degree value?}
\usage{
as_cardinal_degrees(x, winds = 8)
}
\arguments{
\item{x}{A numeric vector with degrees.}

\item{winds}{Number of winds to use for results (4, 8, or 16).}
}
\value{
A named numeric vector with cardinal degrees and wind names.
}
\description{
What are the closest cardinal degrees to a numeric degree value?
}
