% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sf.R
\name{as_crs}
\alias{as_crs}
\title{Coerce a sf, sfc, or bbox object to a coordinate reference system}
\usage{
as_crs(x, input = TRUE)
}
\arguments{
\item{x}{A sf, sfc, or bbox object to coerce into a CRS. Character strings
starting with EPSG: or ESRI: are returned as is. Numeric vectors are
returned as is without checking.}

\item{input}{If \code{TRUE} (default), return only the "input" component of the
crs object. If \code{FALSE}, return the full crs object.}
}
\description{
This function should be updated to support stars objects.
}
