% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numbered_labels.R
\name{as_numbered_labels}
\alias{as_numbered_labels}
\title{Convert a numeric vector to a vector of numbered labels.}
\usage{
as_numbered_labels(
  x,
  labels = "arabic",
  start = NULL,
  suffix = NULL,
  base = 26,
  col = "num_label",
  quiet = TRUE,
  pad = NULL,
  call = parent.frame()
)
}
\arguments{
\item{x}{An integer or other vector or a data.frame. An integer vector or
integer column is used as the number that is converted based on the label
style. If x is not an integer or data.frame with an integer column, the
numbering is created based on \code{\link[=seq_along]{seq_along()}}.}

\item{start}{Starting number or value. Letters are supported if label style
is "alph", "Alph", "alpha", or  "Alpha" and Roman numerals are supported if
label is "roman" or "Roman".}

\item{suffix}{Suffix character to follow number labels. For example, if \code{x = 1} and \code{suffix = "."} the returned label would be "1."}

\item{base}{Base used in alphabetical number labels. Highest letter to use
for alphabetical numbers. Single digit letters (A to Z) or numbers 1 to 26
are supported. For example, if base is 3, alphabetical labels for numbers
higher than 3 have the prior value prefixed so 3 would be "C" and 4 would
be "AA". Defaults to 26 which converts 27 to "AA", 53 to "BA", etc.}

\item{col}{Column name to use for added column for number labels when x is a
data.frame. Defaults to "num_label". If col is length 2, the first item in
the vector is assumed to be the column name from the data.frame to use as x
and the second item is used as the column name for the added column with
number labels.}

\item{quiet}{If \code{TRUE}, suppress warnings about creation of NA values through
coercion of object types. Default to \code{TRUE}.}

\item{pad}{If not \code{NULL} use pad as single padding character to ensure a
consistent character width for all number labels.}

\item{call}{Default: \code{\link[=parent.frame]{parent.frame()}}. Passed to input checking functions
to improve error message traceback.}

\item{label}{Label style. Options include "arabic", "alph", "Alph", "alpha",
"Alpha", "roman", or "Roman".}
}
\value{
\itemize{
\item If x is a vector, function returns numeric vector if labels is
"arabic" or a character vector otherwise.
\item If x is a data.frame, \code{\link[=as_numbered_labels]{as_numbered_labels()}} returns a modified data.frame
with an added column with a name matching the second value of the cols
parameter.
}
}
\description{
Convert a numeric vector to a vector of numbered labels.
}
