% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_orientation.R
\name{as_orientation}
\alias{as_orientation}
\title{What is the orientation of a numeric aspect ratio?}
\usage{
as_orientation(x, tolerance = 0.1, cols = c("width", "height"))
}
\arguments{
\item{x}{A numeric vector with an aspect ratio or a data.frame with width and
height column (names matching the cols parameter).}

\item{tolerance}{Positive numeric value above or below 1 used to determine if
an aspect ratio is square, landscape, or portrait.}

\item{cols}{Name of width and height column if x is a data.frame object.}
}
\description{
What is the orientation of a numeric aspect ratio?
}
