% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_type.R
\name{as_roman}
\alias{as_roman}
\title{as.roman with option to suppress warnings for NA coercion}
\usage{
as_roman(x, quiet = TRUE)
}
\arguments{
\item{x}{a numeric or character vector of arabic or roman numerals.}
}
\description{
as.roman with option to suppress warnings for NA coercion
}
