% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_class.R
\name{as_sfc}
\alias{as_sfc}
\title{Coerce a sf object to a sfc object}
\usage{
as_sfc(x)
}
\arguments{
\item{x}{A sf object to coerce.}
}
\description{
If possible, function should be updated to support bbox objects or other
spatial data classes.
}
