% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{file_path}
\alias{file_path}
\title{Construct path to file ignoring NULL values for filename or path}
\usage{
file_path(
  ...,
  filename = NULL,
  path = NULL,
  fsep = .Platform$file.sep,
  allow_null = FALSE
)
}
\arguments{
\item{...}{Additional strings to pass before path and filename.}

\item{filename}{File name. Optional if path is supplied.}

\item{path}{Path name. Optional if filename is supplied.}

\item{fsep}{the path separator to use (assumed to be ASCII).}

\item{allow_null}{If \code{TRUE}, return \code{NULL} if filename and path are \code{NULL}
and no additional strings are provided to .... If \code{FALSE}, stop if filename
and path are \code{NULL} and no additional strings are provided to ...}
}
\description{
A replacement for \code{file.path()}
}
