% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_named.R
\name{has_all_names}
\alias{has_all_names}
\title{Does an object have all of the provided names?}
\usage{
has_all_names(x, name)
}
\arguments{
\item{x}{A data frame or another named object.}

\item{name}{Element name(s) to check.}
}
\description{
Does an object have all of the provided names?
}
