% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_named.R
\name{has_any_names}
\alias{has_any_names}
\title{Is a named list or character vector?}
\usage{
has_any_names(x, name)
}
\arguments{
\item{x}{A data frame or another named object.}

\item{name}{Element name(s) to check.}
}
\description{
Is a named list or character vector?
}
