% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_filetype.R
\name{has_fileext}
\alias{has_fileext}
\title{Does string contain the specified file type or any file extension?}
\usage{
has_fileext(string = NULL, fileext = NULL, ignore.case = FALSE)
}
\arguments{
\item{string}{String to be tested with or without filetype. Defaults to
\code{NULL}.}

\item{fileext}{File type to test against. Optional.}

\item{ignore.case}{If \code{FALSE}, the pattern matching is case sensitive. If
\code{TRUE}, case is ignored.}
}
\description{
Check if string contains any filetype or the provided filetype. If string is
\code{NULL}, returns \code{FALSE}.
}
\seealso{
\code{\link[=is_fileext_path]{is_fileext_path()}}
}
