% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_filetype.R
\name{has_filetype}
\alias{has_filetype}
\title{Does string contain the specified file type or any file extension?}
\usage{
has_filetype(string = NULL, filetype = NULL, ignore.case = FALSE)
}
\arguments{
\item{string}{String to be tested with or without filetype. Defaults to
\code{NULL}.}

\item{filetype}{File type to test against. Optional.}

\item{ignore.case}{If \code{FALSE}, the pattern matching is case sensitive. If
\code{TRUE}, case is ignored.}
}
\description{
Alternate naming convention for \code{\link[=has_fileext]{has_fileext()}}
}
