% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sf.R
\name{has_same_crs}
\alias{has_same_crs}
\title{Do two sf, sfc, or bbox objects use the same coordinate reference system?}
\usage{
has_same_crs(x, y, ...)
}
\arguments{
\item{x, y}{sf, sfc, or bbox objects to be compared.}

\item{...}{Additional parameters passed to \code{\link[=identical]{identical()}}}
}
\description{
Do two sf, sfc, or bbox objects use the same coordinate reference system?
}
