% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R, R/length.R
\name{has_same_len}
\alias{has_same_len}
\title{Do two object have an identical length?}
\usage{
has_same_len(x, y, ...)

has_same_len(x, y, ...)
}
\arguments{
\item{x, y}{Two strings or character vectors to compare.}

\item{...}{Additional parameters passed to \code{\link[=identical]{identical()}}}
}
\description{
Do two object have an identical length?

Do two object have an identical length?
}
