% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numbered_labels.R
\name{int_to_alpha}
\alias{int_to_alpha}
\title{Convert a integer into a corresponding letter or multi-letter string}
\source{
Adapted from the recursive solution provided by G. Grothendieck in \href{https://stackoverflow.com/a/44274075}{a May 31, 2017 StackOverflow answer}.
}
\usage{
int_to_alpha(x, suffix = NULL, base = 26, dict = LETTERS, quiet = TRUE)
}
\arguments{
\item{x}{An integer vector or a vector that can be coerced to an integer
vector}

\item{suffix}{Suffix character to follow alpha character, e.g. if \code{x = 1} and
\code{suffix = "."} the returned label would be "A.". suffix is also used to
separate values when x is greater than base, e.g. \code{x = 27} and \code{suffix = "."} returns "A.A." Defaults to \code{NULL}.}

\item{base}{If base is not numeric, it is converted to an integer with
\code{\link[=alpha_to_int]{alpha_to_int()}}.}

\item{dict}{Character vector to compare to x. Default: LETTERS.}

\item{quiet}{If \code{TRUE}, suppress warnings for introduction of NA values
through coercion.}
}
\value{
An integer vector composed of objects between 1 and 26 with the same
length as x.
}
\description{
Character values in the provided dict (default to letters "A" to "Z") are
passed as is. Non-integer numeric values or characters that are not found in
the provided dict are converting to NA values.
}
