% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_all.R
\name{is_all_na}
\alias{is_all_na}
\title{Are all items in a list or vector NA values?}
\usage{
is_all_na(x)
}
\arguments{
\item{x}{A list or vector to check.}
}
\description{
Are all items in a list or vector NA values?
}
