% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_all.R
\name{is_all_null}
\alias{is_all_null}
\title{Are all items in a list or vector NULL values?}
\usage{
is_all_null(x)
}
\arguments{
\item{x}{A list or vector to check.}
}
\description{
Are all items in a list or vector NULL values?
}
