% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_any.R
\name{is_any_null}
\alias{is_any_null}
\title{Is any item in a list or vector a NULL value?}
\usage{
is_any_null(x)
}
\arguments{
\item{x}{A list or vector to check.}
}
\description{
Is any item in a list or vector a NULL value?
}
