% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied.R
\name{is_blank}
\alias{is_blank}
\title{Test if a character vector consists of blank strings}
\source{
Adapted from \code{\link[xfun:is_blank]{xfun::is_blank()}} in the
\href{https://yihui.org/xfun/}{xfun} package.
}
\usage{
is_blank(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
\code{TRUE} if the input is blank, or \code{FALSE} otherwise.
}
\description{
Return \code{TRUE} if all elements of a character vector are blank (white
spaces or empty strings).
}
\examples{
is_blank("")
is_blank("abc")
is_blank(c("", "  ", "\n\t"))
is_blank(c("", " ", "abc"))
}
\author{
Yihui Xie \email{xie@yihui.name}
(\href{https://orcid.org/0000-0003-0645-5666}{ORCID})
}
