% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_esri_url}
\alias{is_esri_url}
\title{Is a character vector an ArcGIS MapServer or FeatureServer URL?}
\usage{
is_esri_url(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Is a character vector an ArcGIS MapServer or FeatureServer URL?
}
