% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_path.R
\name{is_excel_path}
\alias{is_excel_path}
\title{Is this a Excel file path or url?}
\usage{
is_excel_path(x, ignore.case = TRUE)
}
\arguments{
\item{x}{A character vector to check.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}
}
\description{
Is this a Excel file path or url?
}
