% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{is_file}
\alias{is_file}
\alias{is_dir}
\title{Is x a file or directory?}
\usage{
is_file(x, include_dirs = FALSE, use_names = FALSE)

is_dir(x, use_names = FALSE)
}
\arguments{
\item{include_dirs}{If \code{TRUE}, return \code{TRUE} for any value of x that is a
directory path. If \code{FALSE} (default), return \code{FALSE} for directory paths.}

\item{use_names}{If \code{TRUE}, return a logical vector where the names match the
values of the input vector x. Defaults to \code{FALSE}.}
}
\description{
\code{\link[=is_file]{is_file()}} is a wrapper for \code{\link[base:files]{base::file.exists()}} that allows the exclusion
of directories and returning named vectors. \code{\link[=is_dir]{is_dir()}} is a wrapper for
\code{\link[base:files2]{base::dir.exists()}} that supports vector inputs rather than single strings.
character(0) inputs return \code{FALSE}.
}
