% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_path.R
\name{is_fileext_path}
\alias{is_fileext_path}
\title{Is this a file path or url ending in the specified file extension?}
\usage{
is_fileext_path(x, fileext, ignore.case = TRUE)
}
\description{
Is this a file path or url ending in the specified file extension?
}
