% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_path.R
\name{is_filetype_path}
\alias{is_filetype_path}
\title{Is this a file path or url ending in the specified filetype?}
\usage{
is_filetype_path(x, filetype, ignore.case = TRUE)
}
\arguments{
\item{x}{A character vector to check.}

\item{filetype}{A file extension (or multiple file extensions) to compare to
x. Required.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}
}
\description{
Is this a file path or url ending in the specified filetype?
}
