% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_list.R
\name{is_gg_list}
\alias{is_gg_list}
\title{Do all items in this list inherit the gg class?}
\usage{
is_gg_list(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Do all items in this list inherit the gg class?
}
