% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_gist_url}
\alias{is_gist_url}
\title{Is a character vector a URL for a GitHub Gist?}
\usage{
is_gist_url(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Is a character vector a URL for a GitHub Gist?
}
