% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_gmap_url}
\alias{is_gmap_url}
\title{Is a character vector a Google Maps URL?}
\usage{
is_gmap_url(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Is a character vector a Google Maps URL?
}
