% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_gsheet_url}
\alias{is_gsheet_url}
\title{Is a character vector a Google Sheets URL?}
\usage{
is_gsheet_url(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Is a character vector a Google Sheets URL?
}
