% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_list.R
\name{is_list_of}
\alias{is_list_of}
\title{Do all items in this list inherit the provided class?}
\usage{
is_list_of(x, what = NULL)
}
\arguments{
\item{x}{Object to be tested.}

\item{what}{A character vector naming classes.}
}
\description{
Do all items in this list inherit the provided class?
}
