% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sf.R
\name{is_lonlat}
\alias{is_lonlat}
\title{Is this sf, sfc, or bbox class object in a geographic coordinate system?}
\usage{
is_lonlat(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
This function currently only checks for "EPSG:4326" and "EPSG:4269" CRS.
}
