% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sf.R
\name{is_sf_ext}
\alias{is_sf_ext}
\title{Is this a sf, sfc, or bbox class object?}
\usage{
is_sf_ext(x, ext = TRUE)
}
\arguments{
\item{x}{Object to be tested.}

\item{ext}{If \code{TRUE}, return \code{TRUE} is x is a sf, sfc, or bbox object. If
\code{FALSE}, only check if x is an sf object. If ext is a character object, it
is passed to the what parameter of \code{\link[=inherits]{inherits()}} with sf.}
}
\description{
Is this a sf, sfc, or bbox class object?
}
