% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_list.R
\name{is_sf_list}
\alias{is_sf_list}
\title{Do all items in this list inherit the sf class?}
\usage{
is_sf_list(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Do all items in this list inherit the sf class?
}
