% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_class.R, R/is_misc.R
\name{is_units}
\alias{is_units}
\title{Is this a units class object?}
\usage{
is_units(x)

is_units(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Is this a units class object?

Is this a units class object?
}
