% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied.R
\name{map_chr}
\alias{map_chr}
\title{Apply a function to each element of a vector.}
\source{
\href{https://github.com/wch/staticimports/blob/main/inst/staticexports/purrr.R}{purr-like functions} in \href{https://wch.github.io/staticimports/}{staticimports} package
}
\usage{
map_chr(.x, .f, ...)
}
\description{
Apply a function to each element of a vector.
}
\author{
Winston Chang \email{winston@stdout.org}
}
