% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\name{plural_words}
\alias{plural_words}
\title{Simple helper for pluralizing words}
\usage{
plural_words(
  words,
  n = 1,
  suffix = "s",
  before = "",
  after = "",
  replacement = NULL
)
}
\description{
Simple helper for pluralizing words
}
