% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numbered_labels.R
\name{roman_to_int}
\alias{roman_to_int}
\title{Convert a Roman numeral character object into a corresponding integer}
\usage{
roman_to_int(x, quiet = TRUE)
}
\arguments{
\item{x}{An integer vector or a character vector with characters representing
Roman numerals.}

\item{quiet}{If \code{TRUE}, suppress warnings for introduction of NA values
through coercion.}
}
\description{
Integers and NA objects are passed as is. Double numeric objects or
characters with no corresponding Roman numeral are converting to NA values.
}
