% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numbered_labels.R
\name{set_start_number}
\alias{set_start_number}
\title{Set start number for numeric vector x}
\usage{
set_start_number(x, start = NULL, labels = "arabic")
}
\arguments{
\item{x}{An integer or other vector or a data.frame. An integer vector or
integer column is used as the number that is converted based on the label
style. If x is not an integer or data.frame with an integer column, the
numbering is created based on \code{\link[=seq_along]{seq_along()}}.}

\item{start}{Starting number or value. Letters are supported if label style
is "alph", "Alph", "alpha", or  "Alpha" and Roman numerals are supported if
label is "roman" or "Roman".}
}
\description{
Helper for \code{\link[=as_numbered_labels]{as_numbered_labels()}}.
}
