% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{static_check}
\alias{static_check}
\alias{static_check_if}
\alias{static_check_character}
\alias{static_check_numeric}
\alias{static_check_nchar}
\alias{static_check_name}
\title{Static checks}
\usage{
static_check_if(condition, message = NULL, call = parent.frame())

static_check_character(x, call = parent.frame())

static_check_numeric(x, call = parent.frame())

static_check_nchar(x, n = 1, ..., call = parent.frame())

static_check_name(x, name = NULL, call = parent.frame())
}
\arguments{
\item{condition}{Condition to check.}

\item{message}{Message to pass to stop if condition is FALSE. Defaults to
\code{NULL}.}

\item{call}{Call passed to stop. Defaults to parent.frame()}
}
\description{
Static checks
}
\keyword{internal}
