% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{str_n_freq}
\alias{str_n_freq}
\title{Get the n most frequent or least frequent appearing values in a vector}
\usage{
str_n_freq(string = NULL, n = NULL, decreasing = TRUE)
}
\arguments{
\item{string}{A character vector.}

\item{n}{The number of values to return based on frequency of appearance.
Defaults to \code{NULL} which returns all unique values from x.}

\item{decreasing}{Passed to \code{\link[=sort]{sort()}}}
}
\description{
This function does not address ties in frequency.
}
