% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied.R
\name{tosentence}
\alias{tosentence}
\title{Convert to a common sentence case}
\source{
\href{https://statisticsglobe.com/r-capitalize-first-letter-of-character-string-containing-multiple-words}{Statistics Globe}
}
\usage{
tosentence(x)
}
\arguments{
\item{x}{a character vector, or an object that can be coerced to character by
\code{\link[=as.character]{as.character()}}.}
}
\description{
Convert to a common sentence case
}
\author{
Joachim Schork \email{info@joachimschork.com}
}
