% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{underscore}
\alias{underscore}
\title{Replace spaces in string with underscores}
\usage{
underscore(string)
}
\arguments{
\item{string}{String to transform by replacing spaces with underscores.}
}
\description{
Replace spaces in string with underscores
}
