% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_crs.R
\name{as_crs}
\alias{as_crs}
\alias{is_lonlat_crs}
\title{Coerce a sf, sfc, or bbox object to a coordinate reference system}
\usage{
as_crs(x, input = TRUE)

is_lonlat_crs(x, crs = c("EPSG:4326", "EPSG:4269"))
}
\arguments{
\item{x}{A sf, sfc, or bbox object to coerce into a CRS.}

\item{input}{If \code{TRUE} (default), return only the "input" component of the
crs object. If \code{FALSE}, return the full crs object.}

\item{crs}{For is_lonlat_crs, coordinate reference system to use as lonlat crs.}
}
\description{
This function should be updated to support stars and terra objects.
}
