% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_type.R
\name{as_integer}
\alias{as_integer}
\title{as.integer with option to suppress warnings for NA coercion}
\usage{
as_integer(x, quiet = TRUE)
}
\arguments{
\item{x}{object to be coerced or tested.}
}
\description{
as.integer with option to suppress warnings for NA coercion
}
