% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied.R
\name{combine_words}
\alias{combine_words}
\title{Combine multiple words into a single string}
\source{
Adapted from \code{\link[knitr:combine_words]{knitr::combine_words()}} in the
\href{https://yihui.org/knitr/}{knitr} package.
}
\usage{
combine_words(
  words,
  sep = ", ",
  and = " and ",
  before = "",
  after = before,
  oxford_comma = TRUE
)
}
\arguments{
\item{words}{A character vector.}

\item{sep}{Separator to be inserted between words.}

\item{and}{Character string to be prepended to the last word.}

\item{before, after}{A character string to be added before/after each word.}

\item{oxford_comma}{Whether to insert the separator between the last two
elements in the list.}
}
\description{
Combine multiple words into a single string
}
\author{
Yihui Xie \email{xie@yihui.name}
(\href{https://orcid.org/0000-0003-0645-5666}{ORCID})
}
