% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_filetype.R
\name{has_filetype}
\alias{has_filetype}
\title{Does string contain the specified file type or any file extension?}
\usage{
has_filetype(string = NULL, filetype = NULL, ignore.case = FALSE)
}
\description{
Alternate naming convention for \code{\link[=has_fileext]{has_fileext()}}
}
