% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{has_len_between}
\alias{has_len_between}
\title{Is the length of x between two values?}
\usage{
has_len_between(x, left, right)
}
\arguments{
\item{x}{Object to check.}

\item{left, right}{Min and max values to check if the length of x is between.}
}
\description{
Is the length of x between two values?
}
