% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_crs.R
\name{has_same_crs}
\alias{has_same_crs}
\title{\itemize{
\item \code{\link[=has_same_crs]{has_same_crs()}}: Do two sf, sfc, or bbox objects use the same coordinate
reference system?
}}
\usage{
has_same_crs(x, y, ...)
}
\arguments{
\item{x, y}{sf, sfc, or bbox objects to be compared.}

\item{...}{Additional parameters passed to \code{\link[=identical]{identical()}} by \code{\link[=has_same_crs]{has_same_crs()}}.}
}
\description{
\itemize{
\item \code{\link[=has_same_crs]{has_same_crs()}}: Do two sf, sfc, or bbox objects use the same coordinate
reference system?
}
}
