% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_all.R
\name{is_all}
\alias{is_all}
\title{Do all items in a list or vector return TRUE from a predicate function?}
\usage{
is_all(x, FUN)
}
\arguments{
\item{x}{A list or vector passed to \code{\link[=vapply]{vapply()}}.}

\item{FUN}{Function passed to FUN parameter of \code{\link[=vapply]{vapply()}}.}
}
\description{
Do all items in a list or vector return TRUE from a predicate function?
}
