% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_any.R
\name{is_any}
\alias{is_any}
\title{Is any item in a list or vector return TRUE from a predicate function?}
\usage{
is_any(x, FUN)
}
\arguments{
\item{x}{A list or vector passed to \code{\link[=vapply]{vapply()}}.}

\item{FUN}{Function passed to FUN parameter of \code{\link[=vapply]{vapply()}}.}
}
\description{
Is any item in a list or vector return TRUE from a predicate function?
}
