% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_any.R
\name{is_any_na}
\alias{is_any_na}
\title{Is any item in a list or vector a NA value?}
\usage{
is_any_na(x)
}
\arguments{
\item{x}{A list or vector to check.}
}
\description{
Is any item in a list or vector a NA value?
}
