% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_path.R
\name{is_fileext_path}
\alias{is_fileext_path}
\title{Is this a file path or url ending in the specified file extension?}
\usage{
is_fileext_path(x, fileext, ignore.case = TRUE)
}
\arguments{
\item{x}{A character vector to check.}

\item{fileext}{A file extension (or multiple file extensions) to compare to
x. Required.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}
}
\description{
Is this a file path or url ending in the specified file extension?
}
