% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_path.R
\name{is_filetype_path}
\alias{is_filetype_path}
\title{Is this a file path or url ending in the specified filetype?}
\usage{
is_filetype_path(x, filetype, ignore.case = TRUE)
}
\description{
Is this a file path or url ending in the specified filetype?
}
